refactor(cli): extract archive helpers and zip command controller
